Set identifiers from external system link.
